Witch.Instances

Documentation

realFloatToRational :: RealFloat s => s -> Either ArithException Rational

overPositive :: (Eq a, Num a, Num b) => (a -> b) -> a -> b

fromDigits :: [Int] -> Int -> (Integer, Integer)

makeRational :: Integer -> Integer -> Rational

fromNonNegativeIntegral :: (Integral s, Num t) => s -> Either ArithException t

maxFloat :: Num a => a

The maximum integral value that can be unambiguously represented as a Float. Equal to 16,777,215.

maxDouble :: Num a => a

The maximum integral value that can be unambiguously represented as a Double. Equal to 9,007,199,254,740,991.

tryEvaluate :: Exception e => a -> Either e a

Orphan instances

From Pico DiffTime
Uses realToFrac.

Instance details

Methods

from :: Pico -> DiffTime

From Pico NominalDiffTime
Uses secondsToNominalDiffTime.

Instance details

Methods

from :: Pico -> NominalDiffTime

From Int16 Int32
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Int32

From Int16 Int64
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Int64

From Int16 Integer
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Integer

From Int16 Double
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Double

From Int16 Float
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Float

From Int16 Int
Uses fromIntegral.

Instance details

Methods

from :: Int16 -> Int

From Int32 Int64
Uses fromIntegral.

Instance details

Methods

from :: Int32 -> Int64

From Int32 Integer
Uses fromIntegral.

Instance details

Methods

from :: Int32 -> Integer

From Int32 Double
Uses fromIntegral.

Instance details

Methods

from :: Int32 -> Double

From Int64 Integer
Uses fromIntegral.

Instance details

Methods

from :: Int64 -> Integer

From Int8 Int16
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Int16

From Int8 Int32
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Int32

From Int8 Int64
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Int64

From Int8 Integer
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Integer

From Int8 Double
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Double

From Int8 Float
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Float

From Int8 Int
Uses fromIntegral.

Instance details

Methods

from :: Int8 -> Int

From Rational UniversalTime
Uses ModJulianDate.

Instance details

Methods

from :: Rational -> UniversalTime

From Rational TimeOfDay
Uses dayFractionToTimeOfDay.

Instance details

Methods

from :: Rational -> TimeOfDay

From Rational Double
Uses fromRational. This necessarily loses some precision.

Instance details

Methods

from :: Rational -> Double

From Rational Float
Uses fromRational. This necessarily loses some precision.

Instance details

Methods

from :: Rational -> Float

From Word16 Int32
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Int32

From Word16 Int64
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Int64

From Word16 Word32
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Word32

From Word16 Word64
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Word64

From Word16 Integer
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Integer

From Word16 Natural
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Natural

From Word16 Double
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Double

From Word16 Float
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Float

From Word16 Int
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Int

From Word16 Word
Uses fromIntegral.

Instance details

Methods

from :: Word16 -> Word

From Word32 Int64
Uses fromIntegral.

Instance details

Methods

from :: Word32 -> Int64

From Word32 Word64
Uses fromIntegral.

Instance details

Methods

from :: Word32 -> Word64

From Word32 Integer
Uses fromIntegral.

Instance details

Methods

from :: Word32 -> Integer

From Word32 Natural
Uses fromIntegral.

Instance details

Methods

from :: Word32 -> Natural

From Word32 Double
Uses fromIntegral.

Instance details

Methods

from :: Word32 -> Double

From Word64 Integer
Uses fromIntegral.

Instance details

Methods

from :: Word64 -> Integer

From Word64 Natural
Uses fromIntegral.

Instance details

Methods

from :: Word64 -> Natural

From Word8 Int16
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Int16

From Word8 Int32
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Int32

From Word8 Int64
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Int64

From Word8 Word16
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Word16

From Word8 Word32
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Word32

From Word8 Word64
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Word64

From Word8 Integer
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Integer

From Word8 Natural
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Natural

From Word8 Double
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Double

From Word8 Float
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Float

From Word8 Int
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Int

From Word8 Word
Uses fromIntegral.

Instance details

Methods

from :: Word8 -> Word

From ByteString ByteString
Uses fromStrict.

Instance details

Methods

from :: ByteString0 -> ByteString

From ByteString ShortByteString
Uses toShort.

Instance details

Methods

from :: ByteString -> ShortByteString

From ByteString ByteString
Uses toStrict.

Instance details

Methods

from :: ByteString -> ByteString0

From ShortByteString ByteString
Uses fromShort.

Instance details

Methods

from :: ShortByteString -> ByteString

From Text Text
Uses fromStrict.

Instance details

Methods

from :: Text0 -> Text

From Text String
Uses unpack.

Instance details

Methods

from :: Text -> String

From Text Text
Uses toStrict.

Instance details

Methods

from :: Text -> Text0

From Text String
Uses unpack.

Instance details

Methods

from :: Text -> String

From CalendarDiffDays CalendarDiffTime
Uses calendarTimeDays.

Instance details

Methods

from :: CalendarDiffDays -> CalendarDiffTime

From Day DayOfWeek
Uses dayOfWeek.

Instance details

Methods

from :: Day -> DayOfWeek

From Day Integer
Uses toModifiedJulianDay.

Instance details

Methods

from :: Day -> Integer

From DiffTime Pico
Uses realToFrac.

Instance details

Methods

from :: DiffTime -> Pico

From DiffTime TimeOfDay
Uses timeToTimeOfDay.

Instance details

Methods

from :: DiffTime -> TimeOfDay

From NominalDiffTime Pico
Uses nominalDiffTimeToSeconds.

Instance details

Methods

from :: NominalDiffTime -> Pico

From NominalDiffTime CalendarDiffTime
Uses calendarTimeTime.

Instance details

Methods

from :: NominalDiffTime -> CalendarDiffTime

From POSIXTime UTCTime
Uses posixSecondsToUTCTime.

Instance details

Methods

from :: POSIXTime -> UTCTime

From SystemTime AbsoluteTime
Uses systemToTAITime.

Instance details

Methods

from :: SystemTime -> AbsoluteTime

From SystemTime POSIXTime
Uses systemToPOSIXTime.

Instance details

Methods

from :: SystemTime -> POSIXTime

From SystemTime UTCTime
Uses systemToUTCTime.

Instance details

Methods

from :: SystemTime -> UTCTime

From UTCTime POSIXTime
Uses utcTimeToPOSIXSeconds.

Instance details

Methods

from :: UTCTime -> POSIXTime

From UTCTime SystemTime
Uses utcToSystemTime.

Instance details

Methods

from :: UTCTime -> SystemTime

From UniversalTime Rational
Uses getModJulianDate.

Instance details

Methods

from :: UniversalTime -> Rational

From TimeOfDay Rational
Uses timeOfDayToDayFraction.

Instance details

Methods

from :: TimeOfDay -> Rational

From TimeOfDay DiffTime
Uses timeOfDayToTime.

Instance details

Methods

from :: TimeOfDay -> DiffTime

From ZonedTime UTCTime
Uses zonedTimeToUTC.

Instance details

Methods

from :: ZonedTime -> UTCTime

From String Text
Uses pack. Some Char values cannot be represented in Text and will be replaced with '\xFFFD'.

Instance details

Methods

from :: String -> Text

From String Text
Uses pack. Some Char values cannot be represented in Text and will be replaced with '\xFFFD'.

Instance details

Methods

from :: String -> Text

From Integer Day
Uses ModifiedJulianDay.

Instance details

Methods

from :: Integer -> Day

From Natural Integer
Uses fromIntegral.

Instance details

Methods

from :: Natural -> Integer

From Double Float
Uses double2Float. This necessarily loses some precision.

Instance details

Methods

from :: Double -> Float

From Float Double
Uses float2Double.

Instance details

Methods

from :: Float -> Double

From Int Int64
Uses fromIntegral.

Instance details

Methods

from :: Int -> Int64

From Int Integer
Uses fromIntegral.

Instance details

Methods

from :: Int -> Integer

From Word Word64
Uses fromIntegral.

Instance details

Methods

from :: Word -> Word64

From Word Integer
Uses fromIntegral.

Instance details

Methods

from :: Word -> Integer

From Word Natural
Uses fromIntegral.

Instance details

Methods

from :: Word -> Natural

From a a
Uses id.

Instance details

Methods

from :: a -> a

TryFrom Int16 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Int8) Int8

TryFrom Int16 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Word16) Word16

TryFrom Int16 Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Word32) Word32

TryFrom Int16 Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Word64) Word64

TryFrom Int16 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Word8) Word8

TryFrom Int16 Natural
Uses fromIntegral when the input is not negative.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Natural) Natural

TryFrom Int16 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int16 -> Either (TryFromException Int16 Word) Word

TryFrom Int32 Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Int16) Int16

TryFrom Int32 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Int8) Int8

TryFrom Int32 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Word16) Word16

TryFrom Int32 Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Word32) Word32

TryFrom Int32 Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Word64) Word64

TryFrom Int32 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Word8) Word8

TryFrom Int32 Natural
Uses fromIntegral when the input is not negative.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Natural) Natural

TryFrom Int32 Float
Uses fromIntegral when the input is between -16,777,215 and 16,777,215 inclusive.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Float) Float

TryFrom Int32 Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Int) Int

TryFrom Int32 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int32 -> Either (TryFromException Int32 Word) Word

TryFrom Int64 Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Int16) Int16

TryFrom Int64 Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Int32) Int32

TryFrom Int64 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Int8) Int8

TryFrom Int64 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Word16) Word16

TryFrom Int64 Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Word32) Word32

TryFrom Int64 Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Word64) Word64

TryFrom Int64 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Word8) Word8

TryFrom Int64 Natural
Uses fromIntegral when the input is not negative.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Natural) Natural

TryFrom Int64 Double
Uses fromIntegral when the input is between -9,007,199,254,740,991 and 9,007,199,254,740,991 inclusive.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Double) Double

TryFrom Int64 Float
Uses fromIntegral when the input is between -16,777,215 and 16,777,215 inclusive.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Float) Float

TryFrom Int64 Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Int) Int

TryFrom Int64 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int64 -> Either (TryFromException Int64 Word) Word

TryFrom Int8 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Word16) Word16

TryFrom Int8 Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Word32) Word32

TryFrom Int8 Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Word64) Word64

TryFrom Int8 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Word8) Word8

TryFrom Int8 Natural
Uses fromIntegral when the input is not negative.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Natural) Natural

TryFrom Int8 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int8 -> Either (TryFromException Int8 Word) Word

TryFrom Word16 Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word16 -> Either (TryFromException Word16 Int16) Int16

TryFrom Word16 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word16 -> Either (TryFromException Word16 Int8) Int8

TryFrom Word16 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word16 -> Either (TryFromException Word16 Word8) Word8

TryFrom Word32 Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Int16) Int16

TryFrom Word32 Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Int32) Int32

TryFrom Word32 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Int8) Int8

TryFrom Word32 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Word16) Word16

TryFrom Word32 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Word8) Word8

TryFrom Word32 Float
Uses fromIntegral when the input is less than or equal to 16,777,215.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Float) Float

TryFrom Word32 Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Int) Int

TryFrom Word32 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word32 -> Either (TryFromException Word32 Word) Word

TryFrom Word64 Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Int16) Int16

TryFrom Word64 Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Int32) Int32

TryFrom Word64 Int64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Int64) Int64

TryFrom Word64 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Int8) Int8

TryFrom Word64 Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Word16) Word16

TryFrom Word64 Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Word32) Word32

TryFrom Word64 Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Word8) Word8

TryFrom Word64 Double
Uses fromIntegral when the input is less than or equal to 9,007,199,254,740,991.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Double) Double

TryFrom Word64 Float
Uses fromIntegral when the input is less than or equal to 16,777,215.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Float) Float

TryFrom Word64 Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Int) Int

TryFrom Word64 Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word64 -> Either (TryFromException Word64 Word) Word

TryFrom Word8 Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word8 -> Either (TryFromException Word8 Int8) Int8

TryFrom Integer Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Int16) Int16

TryFrom Integer Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Int32) Int32

TryFrom Integer Int64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Int64) Int64

TryFrom Integer Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Int8) Int8

TryFrom Integer Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Word16) Word16

TryFrom Integer Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Word32) Word32

TryFrom Integer Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Word64) Word64

TryFrom Integer Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Word8) Word8

TryFrom Integer Natural
Uses fromInteger when the input is not negative.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Natural) Natural

TryFrom Integer Double
Uses fromIntegral when the input is between -9,007,199,254,740,991 and 9,007,199,254,740,991 inclusive.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Double) Double

TryFrom Integer Float
Uses fromIntegral when the input is between -16,777,215 and 16,777,215 inclusive.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Float) Float

TryFrom Integer Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Int) Int

TryFrom Integer Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Integer -> Either (TryFromException Integer Word) Word

TryFrom Natural Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Int16) Int16

TryFrom Natural Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Int32) Int32

TryFrom Natural Int64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Int64) Int64

TryFrom Natural Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Int8) Int8

TryFrom Natural Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Word16) Word16

TryFrom Natural Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Word32) Word32

TryFrom Natural Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Word64) Word64

TryFrom Natural Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Word8) Word8

TryFrom Natural Double
Uses fromIntegral when the input is less than or equal to 9,007,199,254,740,991.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Double) Double

TryFrom Natural Float
Uses fromIntegral when the input is less than or equal to 16,777,215.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Float) Float

TryFrom Natural Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Int) Int

TryFrom Natural Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Natural -> Either (TryFromException Natural Word) Word

TryFrom Double Int16
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Int16) Int16

TryFrom Double Int32
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Int32) Int32

TryFrom Double Int64
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Int64) Int64

TryFrom Double Int8
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Int8) Int8

TryFrom Double Rational
Uses floatToDigits when the input is not NaN or infinity.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Rational) Rational

TryFrom Double Word16
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Word16) Word16

TryFrom Double Word32
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Word32) Word32

TryFrom Double Word64
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Word64) Word64

TryFrom Double Word8
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Word8) Word8

TryFrom Double Integer
Converts via Rational when the input is between -9,007,199,254,740,991 and 9,007,199,254,740,991 inclusive.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Integer) Integer

TryFrom Double Natural
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Natural) Natural

TryFrom Double Int
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Int) Int

TryFrom Double Word
Converts via Integer.

Instance details

Methods

tryFrom :: Double -> Either (TryFromException Double Word) Word

TryFrom Float Int16
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Int16) Int16

TryFrom Float Int32
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Int32) Int32

TryFrom Float Int64
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Int64) Int64

TryFrom Float Int8
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Int8) Int8

TryFrom Float Rational
Uses floatToDigits when the input is not NaN or infinity.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Rational) Rational

TryFrom Float Word16
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Word16) Word16

TryFrom Float Word32
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Word32) Word32

TryFrom Float Word64
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Word64) Word64

TryFrom Float Word8
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Word8) Word8

TryFrom Float Integer
Converts via Rational when the input is between -16,777,215 and 16,777,215 inclusive.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Integer) Integer

TryFrom Float Natural
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Natural) Natural

TryFrom Float Int
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Int) Int

TryFrom Float Word
Converts via Integer.

Instance details

Methods

tryFrom :: Float -> Either (TryFromException Float Word) Word

TryFrom Int Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Int16) Int16

TryFrom Int Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Int32) Int32

TryFrom Int Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Int8) Int8

TryFrom Int Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Word16) Word16

TryFrom Int Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Word32) Word32

TryFrom Int Word64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Word64) Word64

TryFrom Int Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Word8) Word8

TryFrom Int Natural
Uses fromIntegral when the input is not negative.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Natural) Natural

TryFrom Int Double
Uses fromIntegral when the input is between -9,007,199,254,740,991 and 9,007,199,254,740,991 inclusive.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Double) Double

TryFrom Int Float
Uses fromIntegral when the input is between -16,777,215 and 16,777,215 inclusive.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Float) Float

TryFrom Int Word
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Int -> Either (TryFromException Int Word) Word

TryFrom Word Int16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Int16) Int16

TryFrom Word Int32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Int32) Int32

TryFrom Word Int64
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Int64) Int64

TryFrom Word Int8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Int8) Int8

TryFrom Word Word16
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Word16) Word16

TryFrom Word Word32
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Word32) Word32

TryFrom Word Word8
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Word8) Word8

TryFrom Word Double
Uses fromIntegral when the input is less than or equal to 9,007,199,254,740,991.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Double) Double

TryFrom Word Float
Uses fromIntegral when the input is less than or equal to 16,777,215.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Float) Float

TryFrom Word Int
Uses toIntegralSized.

Instance details

Methods

tryFrom :: Word -> Either (TryFromException Word Int) Int

From ByteString [Word8]
Uses unpack.

Instance details

Methods

from :: ByteString -> [Word8]

From ByteString [Word8]
Uses unpack.

Instance details

Methods

from :: ByteString -> [Word8]

From ShortByteString [Word8]
Uses unpack.

Instance details

Methods

from :: ShortByteString -> [Word8]

From IntSet [Int]
Uses toAscList.

Instance details

Methods

from :: IntSet -> [Int]

From Text (UTF_16BE ByteString)
Uses encodeUtf16BE.

Instance details

Methods

from :: Text -> UTF_16BE ByteString

From Text (UTF_16BE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_16BE ByteString

From Text (UTF_16LE ByteString)
Uses encodeUtf16LE.

Instance details

Methods

from :: Text -> UTF_16LE ByteString

From Text (UTF_16LE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_16LE ByteString

From Text (UTF_32BE ByteString)
Uses encodeUtf32BE.

Instance details

Methods

from :: Text -> UTF_32BE ByteString

From Text (UTF_32BE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_32BE ByteString

From Text (UTF_32LE ByteString)
Uses encodeUtf32LE.

Instance details

Methods

from :: Text -> UTF_32LE ByteString

From Text (UTF_32LE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_32LE ByteString

From Text (UTF_8 ByteString)
Uses encodeUtf8.

Instance details

Methods

from :: Text -> UTF_8 ByteString

From Text (UTF_8 ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_8 ByteString

From Text (UTF_16BE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_16BE ByteString

From Text (UTF_16BE ByteString)
Uses encodeUtf16BE.

Instance details

Methods

from :: Text -> UTF_16BE ByteString

From Text (UTF_16LE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_16LE ByteString

From Text (UTF_16LE ByteString)
Uses encodeUtf16LE.

Instance details

Methods

from :: Text -> UTF_16LE ByteString

From Text (UTF_32BE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_32BE ByteString

From Text (UTF_32BE ByteString)
Uses encodeUtf32BE.

Instance details

Methods

from :: Text -> UTF_32BE ByteString

From Text (UTF_32LE ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_32LE ByteString

From Text (UTF_32LE ByteString)
Uses encodeUtf32LE.

Instance details

Methods

from :: Text -> UTF_32LE ByteString

From Text (UTF_8 ByteString)
Converts via ByteString.

Instance details

Methods

from :: Text -> UTF_8 ByteString

From Text (UTF_8 ByteString)
Uses encodeUtf8.

Instance details

Methods

from :: Text -> UTF_8 ByteString

From String (UTF_16BE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_16BE ByteString

From String (UTF_16BE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_16BE ByteString

From String (UTF_16LE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_16LE ByteString

From String (UTF_16LE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_16LE ByteString

From String (UTF_32BE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_32BE ByteString

From String (UTF_32BE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_32BE ByteString

From String (UTF_32LE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_32LE ByteString

From String (UTF_32LE ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_32LE ByteString

From String (UTF_8 ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_8 ByteString

From String (UTF_8 ByteString)
Converts via Text.

Instance details

Methods

from :: String -> UTF_8 ByteString

Num a => From a (Complex a)
Uses (:+) with an imaginary part of 0.

Instance details

Methods

from :: a -> Complex a

Integral a => From a (Ratio a)
Uses (%) with a denominator of 1.

Instance details

Methods

from :: a -> Ratio a

TryFrom Text (ISO_8859_1 ByteString)
Converts via String.

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

TryFrom Text (ISO_8859_1 ByteString)
Converts via String.

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

TryFrom Text (ISO_8859_1 ByteString)
Converts via String.

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

TryFrom Text (ISO_8859_1 ByteString)
Converts via String.

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

TryFrom String (ISO_8859_1 ByteString)
Uses pack when each character isLatin1.

Instance details

Methods

tryFrom :: String -> Either (TryFromException String (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

TryFrom String (ISO_8859_1 ByteString)
Uses pack when each character isLatin1.

Instance details

Methods

tryFrom :: String -> Either (TryFromException String (ISO_8859_1 ByteString)) (ISO_8859_1 ByteString)

From Integer (Fixed a)
Uses MkFixed. This means from @Integer @Centi 2 is 0.02 rather than 2.00.

Instance details

Methods

from :: Integer -> Fixed a

HasResolution a => TryFrom Rational (Fixed a)
Uses fromRational as long as there isn't a loss of precision.

Instance details

Methods

tryFrom :: Rational -> Either (TryFromException Rational (Fixed a)) (Fixed a)

From a (Tagged t a)
Uses coerce. Essentially the same as Tagged.

Instance details

Methods

from :: a -> Tagged t a

From (ISO_8859_1 ByteString) Text
Uses decodeLatin1.

Instance details

Methods

from :: ISO_8859_1 ByteString -> Text

From (ISO_8859_1 ByteString) Text
Converts via Text.

Instance details

Methods

from :: ISO_8859_1 ByteString -> Text

From (ISO_8859_1 ByteString) String
Converts via Text.

Instance details

Methods

from :: ISO_8859_1 ByteString -> String

From (ISO_8859_1 ByteString) Text
Converts via Text.

Instance details

Methods

from :: ISO_8859_1 ByteString -> Text

From (ISO_8859_1 ByteString) Text
Uses decodeLatin1.

Instance details

Methods

from :: ISO_8859_1 ByteString -> Text

From (ISO_8859_1 ByteString) String
Converts via Text.

Instance details

Methods

from :: ISO_8859_1 ByteString -> String

From [Word8] ByteString
Uses pack.

Instance details

Methods

from :: [Word8] -> ByteString

From [Word8] ByteString
Uses pack.

Instance details

Methods

from :: [Word8] -> ByteString

From [Word8] ShortByteString
Uses pack.

Instance details

Methods

from :: [Word8] -> ShortByteString

From [Int] IntSet
Uses fromList.

Instance details

Methods

from :: [Int] -> IntSet

(Eq a, Num a) => TryFrom (Complex a) a
Uses realPart when the imaginary part is 0.

Instance details

Methods

tryFrom :: Complex a -> Either (TryFromException (Complex a) a) a

(Eq a, Num a) => TryFrom (Ratio a) a
Uses numerator when the denominator is 1.

Instance details

Methods

tryFrom :: Ratio a -> Either (TryFromException (Ratio a) a) a

TryFrom (UTF_16BE ByteString) Text
Uses decodeUtf16BE.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) Text) Text

TryFrom (UTF_16BE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) Text) Text

TryFrom (UTF_16BE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) String) String

TryFrom (UTF_16BE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) Text) Text

TryFrom (UTF_16BE ByteString) Text
Uses decodeUtf16BE.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) Text) Text

TryFrom (UTF_16BE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16BE ByteString -> Either (TryFromException (UTF_16BE ByteString) String) String

TryFrom (UTF_16LE ByteString) Text
Uses decodeUtf16LE.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) Text) Text

TryFrom (UTF_16LE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) Text) Text

TryFrom (UTF_16LE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) String) String

TryFrom (UTF_16LE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) Text) Text

TryFrom (UTF_16LE ByteString) Text
Uses decodeUtf16LE.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) Text) Text

TryFrom (UTF_16LE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_16LE ByteString -> Either (TryFromException (UTF_16LE ByteString) String) String

TryFrom (UTF_32BE ByteString) Text
Uses decodeUtf32BE.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) Text) Text

TryFrom (UTF_32BE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) Text) Text

TryFrom (UTF_32BE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) String) String

TryFrom (UTF_32BE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) Text) Text

TryFrom (UTF_32BE ByteString) Text
Uses decodeUtf32BE.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) Text) Text

TryFrom (UTF_32BE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32BE ByteString -> Either (TryFromException (UTF_32BE ByteString) String) String

TryFrom (UTF_32LE ByteString) Text
Uses decodeUtf32LE.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) Text) Text

TryFrom (UTF_32LE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) Text) Text

TryFrom (UTF_32LE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) String) String

TryFrom (UTF_32LE ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) Text) Text

TryFrom (UTF_32LE ByteString) Text
Uses decodeUtf32LE.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) Text) Text

TryFrom (UTF_32LE ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_32LE ByteString -> Either (TryFromException (UTF_32LE ByteString) String) String

TryFrom (UTF_8 ByteString) Text
Uses decodeUtf8'.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) Text) Text

TryFrom (UTF_8 ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) Text) Text

TryFrom (UTF_8 ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) String) String

TryFrom (UTF_8 ByteString) Text
Converts via Text.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) Text) Text

TryFrom (UTF_8 ByteString) Text
Uses decodeUtf8'.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) Text) Text

TryFrom (UTF_8 ByteString) String
Converts via Text.

Instance details

Methods

tryFrom :: UTF_8 ByteString -> Either (TryFromException (UTF_8 ByteString) String) String

From (NonEmpty a) [a]
Uses toList.

Instance details

Methods

from :: NonEmpty a -> [a]

From (IntMap v) [(Int, v)]
Uses toAscList.

Instance details

Methods

from :: IntMap v -> [(Int, v)]

From (Seq a) [a]
Uses toList.

Instance details

Methods

from :: Seq a -> [a]

From (Set a) [a]
Uses toAscList.

Instance details

Methods

from :: Set a -> [a]

From [(Int, v)] (IntMap v)
Uses fromList. If there are duplicate keys, later values will overwrite earlier ones.

Instance details

Methods

from :: [(Int, v)] -> IntMap v

From [a] (Seq a)
Uses fromList.

Instance details

Methods

from :: [a] -> Seq a

Ord a => From [a] (Set a)
Uses fromList.

Instance details

Methods

from :: [a] -> Set a

TryFrom [a] (NonEmpty a)
Uses nonEmpty.

Instance details

Methods

tryFrom :: [a] -> Either (TryFromException [a] (NonEmpty a)) (NonEmpty a)

Ord k => From [(k, v)] (Map k v)
Uses fromList. If there are duplicate keys, later values will overwrite earlier ones.

Instance details

Methods

from :: [(k, v)] -> Map k v

HasResolution a => From (Fixed a) Rational
Uses toRational.

Instance details

Methods

from :: Fixed a -> Rational

From (Fixed a) Integer
Uses MkFixed. This means from @Centi @Integer 3.00 is 300 rather than 3.

Instance details

Methods

from :: Fixed a -> Integer

From (Map k v) [(k, v)]
Uses toAscList.

Instance details

Methods

from :: Map k v -> [(k, v)]

From (TryFromException source oldTarget) (TryFromException source newTarget)
Uses coerce.

Instance details

Methods

from :: TryFromException source oldTarget -> TryFromException source newTarget

From (Tagged t a) a
Uses coerce. Essentially the same as unTagged.

Instance details

Methods

from :: Tagged t a -> a

From (Tagged t a) (Tagged u a)
Uses coerce. Essentially the same as retag.

Instance details

Methods

from :: Tagged t a -> Tagged u a